refactor(config): split DNS and routing parsers
