+++ crates/honk-config/src/parser/mod.rs
@@ -1,3 +1,6 @@
+mod dns;
+mod routing;
+
 #[cfg(test)]
 mod tests;
 
@@ -6,11 +9,9 @@ use std::path::{Path, PathBuf};
 
 use crate::Config;
 use crate::config::GlobalConfig;
-use crate::dns::DnsConfig;
 use crate::experimental::ExperimentalConfig;
 use crate::group::Group;
 use crate::node::Node;
-use crate::routing::RoutingConfig;
 use crate::subscription::Subscription;
 use regex::Regex;
 #[derive(Debug, Clone)]
@@ -385,8 +386,8 @@ pub fn parse_dae_config(input: &str) -> Result<Config, crate::ConfigError> {
     for section in &sections {
         match section.name.as_str() {
             "global" => config.global = parse_global_section(section)?,
-            "dns" => config.dns = parse_dns_section(section)?,
-            "routing" => config.routing = parse_routing_section(section)?,
+            "dns" => config.dns = dns::parse_section(section)?,
+            "routing" => config.routing = routing::parse_section(section)?,
             "node" => {
                 for node in parse_node_section(section)? {
                     config.nodes.push(node);
@@ -870,621 +871,6 @@ fn parse_global_section(section: &Section) -> Result<GlobalConfig, crate::Config
     Ok(cfg)
 }
 
-fn parse_dns_section(section: &Section) -> Result<DnsConfig, crate::ConfigError> {
-    let dns_subs =
-        split_nested_sections(&section.body, &["upstream", "routing", "fixed_domain_ttl"])?;
-    let mut cfg = DnsConfig::default();
-    let mut saw_upstream = false;
-    let dns_body = dns_subs.first().map(|s| s.body.as_str()).unwrap_or("");
-    let kv = parse_kv_pairs(dns_body);
-    if let Some(bind) = kv.get("bind") {
-        cfg.bind.clone_from(bind);
-        cfg.bind_endpoint()
-            .map_err(|error| crate::ConfigError::Parse(error.to_string()))?;
-    }
-    if kv.contains_key("hosts_file") {
-        return Err(crate::ConfigError::Parse(
-            "dns.hosts_file was removed; use one or more use_host paths".into(),
-        ));
-    }
-    for (key, source) in dns_body.lines().filter_map(parse_kv_pair) {
-        if key == "use_host" {
-            crate::dns::push_host_source(&mut cfg.hosts, source);
-        }
-    }
-    if let Some(value) = kv.get("client_subnet") {
-        cfg.client_subnet.clone_from(value);
-        cfg.client_subnet_mode()
-            .map_err(|error| crate::ConfigError::Parse(error.to_string()))?;
-    }
-
-    if let Some(v) = kv.get("ipversion_prefer") {
-        cfg.strategy = parse_ip_prefer(v);
-    }
-    if let Some(v) = kv.get("optimistic_cache") {
-        cfg.cache.enabled = parse_bool(v);
-    }
-    if let Some(v) = kv.get("optimistic_cache_ttl") {
-        cfg.cache.ttl = v.parse().unwrap_or(60);
-    }
-    if let Some(v) = kv.get("max_cache_size") {
-        cfg.cache.max_size = v.parse().unwrap_or(10000);
-    }
-
-    for sub in dns_subs.iter().skip(1) {
-        match sub.name.as_str() {
-            "upstream" => {
-                if !saw_upstream {
-                    cfg.upstream.clear();
-                    saw_upstream = true;
-                }
-                cfg.upstream.extend(parse_dns_upstreams(&sub.body));
-            }
-            "routing" => {
-                for req_body in extract_nested_all(&sub.body, "request") {
-                    let has_fallback = has_routing_fallback(&req_body);
-                    let request = parse_dns_request_routing(&req_body);
-                    cfg.routing.request.rules.extend(request.rules);
-                    if !has_fallback {
-                        continue;
-                    }
-                    cfg.routing.request.fallback = request.fallback;
-                    // Sync legacy fallback for callers that only look there.
-                    if let crate::dns::DnsRequestAction::Upstream(ref name) =
-                        cfg.routing.request.fallback
-                    {
-                        cfg.routing.fallback = name.clone();
-                    }
-                }
-                for resp_body in extract_nested_all(&sub.body, "response") {
-                    let has_fallback = has_routing_fallback(&resp_body);
-                    let response = parse_dns_response_routing(&resp_body);
-                    cfg.routing.response.rules.extend(response.rules);
-                    if has_fallback {
-                        cfg.routing.response.fallback = response.fallback;
-                    }
-                }
-            }
-            "fixed_domain_ttl" => {
-                cfg.fixed_domain_ttl
-                    .extend(parse_fixed_domain_ttl(&sub.body));
-            }
-            _ => {}
-        }
-    }
-
-    Ok(cfg)
-}
-
-fn parse_dns_upstreams(body: &str) -> Vec<crate::dns::DnsUpstream> {
-    let mut upstreams = Vec::new();
-    for line in body.lines() {
-        let trimmed = line.trim();
-        if trimmed.is_empty() || trimmed.starts_with('#') {
-            continue;
-        }
-        if let Some(pos) = trimmed.find(':') {
-            let name = trimmed[..pos].trim().to_string();
-            let rest = trimmed[pos + 1..].trim();
-            // Optional via-proxy suffix (same line):
-            //   preferred:  name: 'uri' -> proxy
-            //   legacy:     name: 'uri' outbound: proxy
-            let (uri, outbound) = if let Some((left, right)) = rest.split_once("->") {
-                let uri_part = left.trim().trim_matches('\'').trim_matches('"');
-                let outbound_part = right.trim().trim_matches('\'').trim_matches('"');
-                let outbound = if outbound_part.is_empty() {
-                    None
-                } else {
-                    Some(outbound_part.to_string())
-                };
-                (uri_part, outbound)
-            } else if let Some(opos) = rest.find("outbound:") {
-                let uri_part = rest[..opos].trim().trim_matches('\'').trim_matches('"');
-                let outbound_part = rest[opos + 9..].trim().trim_matches('\'').trim_matches('"');
-                (uri_part, Some(outbound_part.to_string()))
-            } else {
-                (rest.trim_matches('\'').trim_matches('"'), None)
-            };
-            let (protocol, address) = parse_upstream_uri(uri);
-            let (address, explicit_sni) = extract_tls_server_name(address);
-            let tls_server_name = explicit_sni.or_else(|| sni_from_upstream_address(&address));
-            upstreams.push(crate::dns::DnsUpstream {
-                name,
-                address,
-                protocol,
-                tls_server_name,
-                outbound,
-            });
-        }
-    }
-    upstreams
-}
-
-fn parse_upstream_uri(uri: &str) -> (crate::types::DnsProtocol, String) {
-    let uri = uri.trim();
-    if let Some(rest) = uri.strip_prefix("tcp+udp://") {
-        (crate::types::DnsProtocol::Udp, rest.to_string())
-    } else if let Some(rest) = uri.strip_prefix("udp+tcp://") {
-        (crate::types::DnsProtocol::Udp, rest.to_string())
-    } else if let Some(rest) = uri.strip_prefix("h3://") {
-        (crate::types::DnsProtocol::H3, rest.to_string())
-    } else if let Some(rest) = uri.strip_prefix("http3://") {
-        (crate::types::DnsProtocol::H3, rest.to_string())
-    } else if let Some(rest) = uri.strip_prefix("quic://") {
-        (crate::types::DnsProtocol::Quic, rest.to_string())
-    } else if let Some(rest) = uri.strip_prefix("https://") {
-        (crate::types::DnsProtocol::Https, rest.to_string())
-    } else if let Some(rest) = uri.strip_prefix("tls://") {
-        (crate::types::DnsProtocol::Tls, rest.to_string())
-    } else if let Some(rest) = uri.strip_prefix("tcp://") {
-        (crate::types::DnsProtocol::Tcp, rest.to_string())
-    } else if let Some(rest) = uri.strip_prefix("udp://") {
-        (crate::types::DnsProtocol::Udp, rest.to_string())
-    } else {
-        (crate::types::DnsProtocol::Udp, uri.to_string())
-    }
-}
-
-/// Derive a TLS SNI hostname from a stripped upstream address.
-///
-/// Returns `None` when the host is a bare IP (no SNI needed / not useful).
-fn sni_from_upstream_address(address: &str) -> Option<String> {
-    let hostport = address.split('/').next().unwrap_or(address);
-    let host = if let Some(rest) = hostport.strip_prefix('[') {
-        rest.split(']').next().unwrap_or(rest)
-    } else {
-        hostport
-            .rsplit_once(':')
-            .map(|(h, p)| {
-                // Only treat as host:port when the suffix is numeric.
-                if p.chars().all(|c| c.is_ascii_digit()) {
-                    h
-                } else {
-                    hostport
-                }
-            })
-            .unwrap_or(hostport)
-    };
-    let host = host.trim();
-    if host.is_empty() {
-        return None;
-    }
-    // Bare IPs do not need (and often cannot use) SNI.
-    if host.parse::<std::net::IpAddr>().is_ok() {
-        return None;
-    }
-    Some(host.to_string())
-}
-
-/// Strip an explicit `tls_server_name=` query parameter from an upstream
-/// address, e.g. `tls://1.1.1.1:853?tls_server_name=cloudflare-dns.com`.
-/// Needed for IP-literal TLS upstreams whose certificate hostname differs
-/// from the dial address. Other query pairs are preserved.
-fn extract_tls_server_name(address: String) -> (String, Option<String>) {
-    let Some(qpos) = address.find('?') else {
-        return (address, None);
-    };
-    let (base, query) = address.split_at(qpos);
-    let mut sni = None;
-    let mut kept = Vec::new();
-    for pair in query[1..].split('&') {
-        if let Some(v) = pair.strip_prefix("tls_server_name=") {
-            let v = v.trim();
-            if !v.is_empty() {
-                sni = Some(v.to_string());
-            }
-        } else if !pair.is_empty() {
-            kept.push(pair);
-        }
-    }
-    let address = if kept.is_empty() {
-        base.to_string()
-    } else {
-        format!("{base}?{}", kept.join("&"))
-    };
-    (address, sni)
-}
-
-/// Parse `fixed_domain_ttl { domain: N ... }` into a HashMap.
-fn parse_fixed_domain_ttl(body: &str) -> std::collections::HashMap<String, u32> {
-    let mut map = std::collections::HashMap::new();
-    for line in body.lines() {
-        let trimmed = line.trim();
-        if trimmed.is_empty() || trimmed.starts_with('#') {
-            continue;
-        }
-        if let Some(pos) = trimmed.find(':') {
-            let key = trimmed[..pos].trim().trim_matches('"').trim_matches('\'');
-            let val = trimmed[pos + 1..].split_whitespace().next().unwrap_or("");
-            if let Ok(n) = val.parse::<u32>() {
-                map.insert(key.to_string(), n);
-            }
-        }
-    }
-    map
-}
-
-/// Parse `routing.request { ... }` block.
-fn parse_dns_request_routing(body: &str) -> crate::dns::DnsRequestRouting {
-    let mut routing = crate::dns::DnsRequestRouting::default();
-
-    for line in body.lines() {
-        let mut trimmed = line.trim();
-        if let Some(pos) = trimmed.find("//") {
-            trimmed = trimmed[..pos].trim();
-        } else if let Some(pos) = trimmed.find('#') {
-            // Only strip # if preceded by space (to avoid stripping domain # itself)
-            if pos > 0 && trimmed.as_bytes()[pos - 1] == b' ' {
-                trimmed = trimmed[..pos].trim();
-            }
-        }
-        if trimmed.is_empty() {
-            continue;
-        }
-
-        if trimmed.starts_with("fallback:") || trimmed.starts_with("default:") {
-            let fb = trimmed.split_once(':').unwrap().1.trim();
-            routing.fallback = crate::dns::DnsRequestAction::parse(fb);
-            continue;
-        }
-
-        if let Some(arrow_pos) = trimmed.find("->") {
-            let left = trimmed[..arrow_pos].trim();
-            let right = trimmed[arrow_pos + 2..].trim();
-            let action = crate::dns::DnsRequestAction::parse(right);
-            let conditions = parse_dns_conditions(left, false);
-            // Skip rules whose conditions were all ignored (e.g. sub()/node()).
-            if !conditions.is_empty() {
-                routing
-                    .rules
-                    .push(crate::dns::DnsRequestRule { conditions, action });
-            }
-        }
-    }
-
-    routing
-}
-
-/// Parse `routing.response { ... }` block.
-fn parse_dns_response_routing(body: &str) -> crate::dns::DnsResponseRouting {
-    let mut routing = crate::dns::DnsResponseRouting::default();
-
-    for line in body.lines() {
-        let mut trimmed = line.trim();
-        if let Some(pos) = trimmed.find("//") {
-            trimmed = trimmed[..pos].trim();
-        } else if let Some(pos) = trimmed.find('#')
-            && pos > 0
-            && trimmed.as_bytes()[pos - 1] == b' '
-        {
-            trimmed = trimmed[..pos].trim();
-        }
-        if trimmed.is_empty() {
-            continue;
-        }
-
-        if trimmed.starts_with("fallback:") || trimmed.starts_with("default:") {
-            let fb = trimmed.split_once(':').unwrap().1.trim();
-            routing.fallback = crate::dns::DnsResponseAction::parse(fb);
-            continue;
-        }
-
-        if let Some(arrow_pos) = trimmed.find("->") {
-            let left = trimmed[..arrow_pos].trim();
-            let right = trimmed[arrow_pos + 2..].trim();
-            let action = crate::dns::DnsResponseAction::parse(right);
-            let conditions = parse_dns_conditions(left, true);
-            // Skip rules whose conditions were all ignored (e.g. sub()/node()).
-            if !conditions.is_empty() {
-                routing
-                    .rules
-                    .push(crate::dns::DnsResponseRule { conditions, action });
-            }
-        }
-    }
-
-    routing
-}
-
-/// Parse a chain of `&&`-separated conditions.
-fn parse_dns_conditions(expr: &str, is_response: bool) -> Vec<crate::dns::DnsCond> {
-    let mut conds = Vec::new();
-    let parts: Vec<&str> = expr.split("&&").map(|s| s.trim()).collect();
-
-    for part in parts {
-        let part = part.trim();
-        if part.is_empty() {
-            continue;
-        }
-        let (not, inner) = if let Some(rest) = part.strip_prefix('!') {
-            (true, rest.trim())
-        } else {
-            (false, part)
-        };
-
-        if let Some(args) = extract_fn_args(inner, "qname") {
-            let matchers = parse_dns_qname_args(&args);
-            conds.push(crate::dns::DnsCond::Qname { not, matchers });
-            continue;
-        }
-
-        if let Some(args) = extract_fn_args(inner, "qtype") {
-            let types: Vec<u16> = args
-                .iter()
-                .filter_map(|a| crate::dns::parse_qtype_token(a))
-                .collect();
-            conds.push(crate::dns::DnsCond::Qtype { not, types });
-            continue;
-        }
-
-        if let Some(cidrs) = extract_fn_args(inner, "sip") {
-            conds.push(crate::dns::DnsCond::Sip { not, cidrs });
-            continue;
-        }
-
-        if is_response {
-            if let Some(args) = extract_fn_args(inner, "upstream") {
-                conds.push(crate::dns::DnsCond::Upstream { not, names: args });
-                continue;
-            }
-            if let Some(args) = extract_fn_args(inner, "ip") {
-                let (cidrs, geoip) = parse_dns_ip_args(&args);
-                conds.push(crate::dns::DnsCond::Ip { not, cidrs, geoip });
-                continue;
-            }
-        }
-
-        // sub() / node() / subnode() — not supported for client DNS, warn
-        if inner.starts_with("sub(") || inner.starts_with("node(") || inner.starts_with("subnode(")
-        {
-            eprintln!(
-                "dns routing: ignoring unsupported function {} (out of scope for client DNS)",
-                inner
-            );
-            continue;
-        }
-
-        // unknown condition function — silently ignored
-    }
-
-    conds
-}
-
-/// Parse qname(args) into a list of domain matchers.
-fn parse_dns_qname_args(args: &[String]) -> Vec<crate::dns::DnsDomainMatcher> {
-    let mut matchers = Vec::new();
-    for a in args {
-        let a = a.trim();
-        if a.is_empty() {
-            continue;
-        }
-        if let Some(v) = strip_tag_arg(a, "geosite:") {
-            matchers.push(crate::dns::DnsDomainMatcher::Geosite(
-                normalize_geosite_code(&v),
-            ));
-        } else if let Some(v) = strip_tag_arg(a, "keyword:") {
-            matchers.push(crate::dns::DnsDomainMatcher::Keyword(v));
-        } else if let Some(v) = strip_tag_arg(a, "full:") {
-            matchers.push(crate::dns::DnsDomainMatcher::Full(v));
-        } else if let Some(v) = strip_tag_arg(a, "regex:") {
-            matchers.push(crate::dns::DnsDomainMatcher::Regex(v));
-        } else if let Some(v) = strip_tag_arg(a, "suffix:") {
-            matchers.push(crate::dns::DnsDomainMatcher::Suffix(v));
-        } else {
-            // Bare argument → suffix (dae compatible)
-            matchers.push(crate::dns::DnsDomainMatcher::Suffix(a.to_string()));
-        }
-    }
-    matchers
-}
-
-/// Parse ip(...) args into (cidrs, geoip_codes).
-fn parse_dns_ip_args(args: &[String]) -> (Vec<String>, Vec<String>) {
-    let mut cidrs = Vec::new();
-    let mut geoip = Vec::new();
-    for a in args {
-        let a = a.trim();
-        if let Some(v) = strip_tag_arg(a, "geoip:") {
-            geoip.push(v.to_lowercase());
-        } else {
-            cidrs.push(a.to_string());
-        }
-    }
-    (cidrs, geoip)
-}
-
-fn split_routing_statements(body: &str) -> Result<Vec<String>, crate::ConfigError> {
-    let mut statements = Vec::new();
-    let mut current = String::new();
-    let mut parenthesis_depth = 0usize;
-
-    for (line_index, line) in body.lines().enumerate() {
-        let mut chunk = String::new();
-        let mut quote = None;
-        let mut escaped = false;
-
-        for ch in line.chars() {
-            if let Some(delimiter) = quote {
-                chunk.push(ch);
-                if escaped {
-                    escaped = false;
-                } else if ch == '\\' {
-                    escaped = true;
-                } else if ch == delimiter {
-                    quote = None;
-                }
-                continue;
-            }
-
-            match ch {
-                '#' => break,
-                '\'' | '"' => {
-                    quote = Some(ch);
-                    chunk.push(ch);
-                }
-                '(' => {
-                    parenthesis_depth += 1;
-                    chunk.push(ch);
-                }
-                ')' => {
-                    if parenthesis_depth == 0 {
-                        return Err(crate::ConfigError::Parse(format!(
-                            "routing line {}: unmatched ')'",
-                            line_index + 1
-                        )));
-                    }
-                    parenthesis_depth -= 1;
-                    chunk.push(ch);
-                }
-                _ => chunk.push(ch),
-            }
-        }
-
-        if quote.is_some() {
-            return Err(crate::ConfigError::Parse(format!(
-                "routing line {}: unterminated quote",
-                line_index + 1
-            )));
-        }
-
-        let chunk = chunk.trim();
-        if !chunk.is_empty() {
-            if !current.is_empty() && !chunk.starts_with([')', ',']) {
-                current.push(' ');
-            }
-            current.push_str(chunk);
-        }
-
-        if parenthesis_depth == 0 && !current.is_empty() {
-            statements.push(std::mem::take(&mut current));
-        }
-    }
-
-    if parenthesis_depth != 0 {
-        return Err(crate::ConfigError::Parse(
-            "routing section: unterminated parenthesized rule".into(),
-        ));
-    }
-
-    Ok(statements)
-}
-
-fn parse_default_outbound(statement: &str) -> Option<String> {
-    ["fallback:", "default:"]
-        .into_iter()
-        .find_map(|prefix| statement.strip_prefix(prefix))
-        .map(str::trim)
-        .map(str::to_owned)
-}
-
-fn parse_routing_rule(
-    statement: String,
-    index: usize,
-) -> Option<(crate::routing::RoutingRule, Option<String>)> {
-    let (left, right) = statement.split_once("->")?;
-    let left = left.trim();
-    let right = right.trim();
-    let (outbound, must) = right.strip_suffix("(must)").map_or_else(
-        || (right.to_owned(), false),
-        |name| (name.trim().to_owned(), true),
-    );
-    let condition = parse_route_condition(left);
-    let is_complex = must || left.split("&&").nth(1).is_some() || condition.needs_complex_display();
-    let rule = crate::routing::RoutingRule {
-        name: format!("rule-{index}"),
-        condition,
-        outbound: crate::routing::RoutingOutbound::Simple(outbound),
-        priority: index as u32,
-        must,
-        mark: 0,
-    };
-
-    Some((rule, is_complex.then_some(statement)))
-}
-
-fn parse_routing_section(section: &Section) -> Result<RoutingConfig, crate::ConfigError> {
-    split_routing_statements(&section.body).map(|statements| {
-        statements
-            .into_iter()
-            .fold(RoutingConfig::default(), |mut config, statement| {
-                if let Some(outbound) = parse_default_outbound(&statement) {
-                    config.default_outbound = outbound;
-                } else if let Some((rule, complex_source)) =
-                    parse_routing_rule(statement, config.rules.len())
-                {
-                    if let Some(source) = complex_source {
-                        config.record_complex_rule_source(rule.name.clone(), source);
-                    }
-                    config.rules.push(rule);
-                }
-                config
-            })
-    })
-}
-
-fn parse_route_matcher(condition: &mut crate::routing::RoutingCondition, matcher: &str) {
-    let (negated, matcher) = matcher
-        .strip_prefix('!')
-        .map_or((false, matcher), |rest| (true, rest.trim()));
-    if matcher.is_empty() {
-        return;
-    }
-
-    let mut target = if negated {
-        condition.not.fields_mut()
-    } else {
-        condition.fields_mut()
-    };
-    if let Some(args) = extract_fn_args(matcher, "pname") {
-        target.process_name.extend(args);
-    } else if let Some(args) = extract_fn_args(matcher, "dip") {
-        parse_ip_args(&args, &mut target);
-    } else if let Some(args) = extract_fn_args(matcher, "sip") {
-        target.source_ip.extend(args);
-    } else if let Some(args) = extract_fn_args(matcher, "domain") {
-        parse_domain_args(&args, &mut target);
-    } else if let Some(args) = extract_fn_args(matcher, "dport") {
-        target.port.extend(args);
-    } else if let Some(args) = extract_fn_args(matcher, "sport") {
-        target.source_port.extend(args);
-    } else if let Some(args) = extract_fn_args(matcher, "l4proto") {
-        target.protocol.extend(args);
-    } else if let Some(args) = extract_fn_args(matcher, "ipversion") {
-        target.ip_version.extend(args);
-    } else if let Some(args) = extract_fn_args(matcher, "mac") {
-        target.mac.extend(args);
-    } else if let Some(args) = extract_fn_args(matcher, "dscp") {
-        target.dscp.extend(args);
-    } else if let Some(value) = strip_tag_arg(matcher, "geosite:") {
-        target.geosite.push(normalize_geosite_code(&value));
-    } else if let Some(value) = strip_tag_arg(matcher, "geoip:") {
-        target.geo_ip.push(normalize_geosite_code(&value));
-    } else if let Some(value) = strip_tag_arg(matcher, "domain:") {
-        target.domain_suffix.push(value);
-    } else if let Some(value) = strip_tag_arg(matcher, "suffix:") {
-        target.domain_suffix.push(value);
-    } else if let Some(value) = strip_tag_arg(matcher, "keyword:") {
-        target.domain_keyword.push(value);
-    } else if let Some(value) = strip_tag_arg(matcher, "full:") {
-        target.domain.push(value);
-    } else if let Some(value) = strip_tag_arg(matcher, "regex:") {
-        target.domain_regex.push(value);
-    }
-}
-
-fn parse_route_condition(expr: &str) -> crate::routing::RoutingCondition {
-    expr.split("&&")
-        .map(str::trim)
-        .filter(|part| !part.is_empty())
-        .fold(
-            crate::routing::RoutingCondition::default(),
-            |mut condition, matcher| {
-                parse_route_matcher(&mut condition, matcher);
-                condition
-            },
-        )
-}
-
 fn extract_fn_args(expr: &str, fn_name: &str) -> Option<Vec<String>> {
     let args = expr
         .strip_prefix(fn_name)?
@@ -1514,39 +900,6 @@ fn normalize_geosite_code(code: &str) -> String {
     // name — remapping it to `-` silently mismatched into a nonexistent
     // category.
     code.trim().to_string()
-}
-
-/// Dispatch `domain(...)` arguments to the correct condition fields.
-/// Supports `suffix:`, `keyword:`, `full:`, `regex:`, and `geosite:`.
-fn parse_domain_args(args: &[String], cond: &mut crate::routing::ConditionFields<'_>) {
-    for a in args {
-        if let Some(v) = strip_tag_arg(a, "geosite:") {
-            cond.geosite.push(normalize_geosite_code(&v));
-        } else if let Some(v) = strip_tag_arg(a, "keyword:") {
-            cond.domain_keyword.push(v);
-        } else if let Some(v) = strip_tag_arg(a, "full:") {
-            cond.domain.push(v);
-        } else if let Some(v) = strip_tag_arg(a, "regex:") {
-            cond.domain_regex.push(v);
-        } else if let Some(v) = strip_tag_arg(a, "suffix:") {
-            cond.domain_suffix.push(v);
-        } else {
-            // Bare domain argument defaults to suffix matching, mirroring dae.
-            cond.domain_suffix.push(a.trim().to_string());
-        }
-    }
-}
-
-/// Dispatch `dip(...)` arguments to the correct condition fields.
-/// Supports `geoip:` and plain CIDRs.
-fn parse_ip_args(args: &[String], cond: &mut crate::routing::ConditionFields<'_>) {
-    for a in args {
-        if let Some(v) = strip_tag_arg(a, "geoip:") {
-            cond.geo_ip.push(normalize_geosite_code(&v));
-        } else {
-            cond.ip.push(a.trim().to_string());
-        }
-    }
 }
 
 fn node_parse_diagnostic(error: &crate::ConfigError) -> String {
